feat: Add API tab to user settings page
