+++ web/src/components/api-service/hooks.ts
@@ -14,7 +14,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { message } from 'antd';
 import { useCallback } from 'react';
 
-export const useOperateApiKey = (dialogId: string, idKey: string) => {
+export const useOperateApiKey = (idKey: string, dialogId?: string) => {
   const { removeToken } = useRemoveNextToken();
   const { createToken, loading: creatingLoading } = useCreateNextToken();
   const { data: tokenList, loading: listLoading } = useFetchTokenList({
@@ -49,7 +49,7 @@ type ChartStatsType = {
 export const useSelectChartStatsList = (): ChartStatsType => {
   const queryClient = useQueryClient();
   const data = queryClient.getQueriesData({ queryKey: ['fetchStats'] });
-  const stats: IStats = data[0][1] as IStats;
+  const stats: IStats = (data.length > 0 ? data[0][1] : {}) as IStats;
 
   return Object.keys(stats).reduce((pre, cur) => {
     const item = stats[cur as keyof IStats];
@@ -77,7 +77,7 @@ const getUrlWithToken = (token: string, from: string = 'chat') => {
   return `${protocol}//${host}/chat/share?shared_id=${token}&from=${from}`;
 };
 
-const useFetchTokenListBeforeOtherStep = (dialogId: string, idKey: string) => {
+const useFetchTokenListBeforeOtherStep = (idKey: string, dialogId?: string) => {
   const { showTokenEmptyError } = useShowTokenEmptyError();
 
   const { data: tokenList, refetch } = useFetchTokenList({ [idKey]: dialogId });
@@ -102,7 +102,7 @@ const useFetchTokenListBeforeOtherStep = (dialogId: string, idKey: string) => {
   };
 };
 
-export const useShowEmbedModal = (dialogId: string, idKey: string) => {
+export const useShowEmbedModal = (idKey: string, dialogId?: string) => {
   const {
     visible: embedVisible,
     hideModal: hideEmbedModal,
@@ -110,8 +110,8 @@ export const useShowEmbedModal = (dialogId: string, idKey: string) => {
   } = useSetModalState();
 
   const { handleOperate, token } = useFetchTokenListBeforeOtherStep(
-    dialogId,
     idKey,
+    dialogId,
   );
 
   const handleShowEmbedModal = useCallback(async () => {
@@ -129,8 +129,8 @@ export const useShowEmbedModal = (dialogId: string, idKey: string) => {
   };
 };
 
-export const usePreviewChat = (dialogId: string, idKey: string) => {
-  const { handleOperate } = useFetchTokenListBeforeOtherStep(dialogId, idKey);
+export const usePreviewChat = (idKey: string, dialogId?: string) => {
+  const { handleOperate } = useFetchTokenListBeforeOtherStep(idKey, dialogId);
 
   const open = useCallback(
     (t: string) => {
