+++ web/src/constants/setting.ts
@@ -5,6 +5,7 @@ export enum UserSettingRouteKey {
   Password = 'password',
   Model = 'model',
   System = 'system',
+  Api = 'api',
   Team = 'team',
   Logout = 'logout',
 }
