+++ web/src/locales/en.ts
@@ -451,6 +451,7 @@ The above is the content you need to summarize.`,
       team: 'Team',
       system: 'System',
       logout: 'Log out',
+      api: 'API',
       username: 'Username',
       usernameMessage: 'Please input your username!',
       photo: 'Your photo',
