+++ web/src/pages/api/index.tsx
@@ -1,0 +1,11 @@
+import ApiContent from '@/components/api-service/chat-overview-modal/api-content';
+
+const ApiPage = () => {
+  return (
+    <div>
+      <ApiContent idKey="dialogId"></ApiContent>
+    </div>
+  );
+};
+
+export default ApiPage;
