+++ web/src/pages/user-setting/setting-api/index.less
@@ -1,0 +1,7 @@
+.apiWrapper {
+  width: 100%;
+
+  div[class^='chartWrapper'] {
+    height: auto !important;
+  }
+}
